Move consumption-template-related API tests
